Move Advance to cpp file
