+++ Source/Microphysics/Morrison/Make.package
@@ -1,5 +1,6 @@
 CEXE_sources += ERF_InitMorrison.cpp
 CEXE_sources += ERF_UpdateMorrison.cpp
+CEXE_sources += ERF_AdvanceMorrison.cpp
 CEXE_sources += ERF_Morrison_Cloud.cpp
 CEXE_sources += ERF_Morrison_IceFall.cpp
 CEXE_sources += ERF_Morrison_Precip.cpp
